move IniConfiguration to core
