+++ src/PrestaShopBundle/Resources/config/services/core/configuration.yml
@@ -1,0 +1,3 @@
+services:
+    prestashop.core.configuration.ini_configuration:
+        class: 'PrestaShop\PrestaShop\Core\Configuration\IniConfiguration'
