+++ src/PrestaShopBundle/Resources/config/services/core_services.yml
@@ -1,2 +1,3 @@
 imports:
     - { resource: core/import.yml }
+    - { resource: core/configuration.yml }
